Improve the Github Actions Test Workflow
The old one was trying to do a bit too much!

This was copied from the Knowledge Repo, just as a starting point.

We are really relying on the Cargo workspace
so that we don't have to manually search in sub-dirs
or anything like that.

This should fmt, clippy and test and crate in the workspace!